requirements.txt
gradio>=4.44.0
google-generativeai>=0.3.0
faiss-cpu>=1.7.0
duckduckgo-search>=3.8.0
arxiv>=1.4.0
PyMuPDF>=1.23.0
PyPDF2>=3.0.0
beautifulsoup4>=4.12.0
requests>=2.28.0
flask>=2.3.0
flask-cors>=4.0.0
werkzeug>=2.3.0
numpy>=1.24.0,<2.0.0
python-dotenv>=1.0.0